[refactor] Router 내에 있는 기본 Layout 함수를 별도의 DefaultLayout 으로 분리
